pyproject.toml
[project]
name = "autogen-llamaindex-ollama-agenitc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "llama-index-llms-ollama>=0.6.2",
    "llama_index.tools.mcp",
    "mcp[cli]>=1.10.1",
    "mlflow>=3.1.1",
    "ollama>=0.5.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
]
